fix: lock authorities and token authorities can unlock fungibles
Port changes from
https://github.com/GalaChain/sdk/pull/455 to v2

Closes https://github.com/GalaChain/sdk/issues/446

When defined, a lockAuthority on a TokenHold can unlock the token
regardless of ownership.

A token authority can unlock a TokenHold on another users balance
regardless of lockAuthority status.

Refactoring or reworking order of unlocks by token authorities
is out of scope for this immediate fix, should be pursued in a
separate issue if needed.