migrate: swaps (#468)
* migrate: swaps

Migrate Token Swap features from assets-chaincode to public sdk

From the beginning, at least the start of my time on
November 1, 2021, GalaChain has planned to support and
included an implementation of peer-to-peer token
swapping.

This feature has been migrated, updated, and
refined through many refactorings and
changes in chaincode architecture. Many hands
contributed to this feature. It is only
now getting migrated to the public sdk for
wider use, due to various competing demands
throughout 2024. Finally, with
this commit, the remainder of the swap
implementation will be provided with the developer sdk.

* fix: copyright statements