test: BatchSubmit does not preserve writes from failed transactions
// Given
Batch Requests 1, 2, 3

// When
BatchRequest1:

read keyA
write keyA

successful

BatchRequest2:

write keyA <-- gets BatchRequest1's changes

error/failure

BatchRequest3:

read keyA <-- should NOT get BatchRequest2's changes

// Then
BatchRequest3 should not get BatchRequest2's changes

However,

```

Objects are passed around by reference. They are never copied.

-Douglas Crockford (2008), JavaScript: The Good Parts, page 22
```

Tracing that aspect of JavaScript in our code:

https://github.com/GalaChain/sdk/blob/main/chaincode/src/contracts/GalaContract.ts#L222-L239

```
      if (GalaChainResponse.isSuccess(response)) {
        aggregatedCache.writes = ctx.stub.getWrites();
        aggregatedCache.deletes = ctx.stub.getDeletes();
        writesCount = Object.keys(aggregatedCache.writes).length;
      } else {
        ctx.stub.setWrites(aggregatedCache.writes);
        ctx.stub.setDeletes(aggregatedCache.deletes);
      }
      ctx.stub.setReads({});
```

Think about this hypothetical pseudo-code:

```
aggregatedCache.writes[keyA] = {reference to valueA}
```

That would mean the following is also true:

```
ctx.stub.writes[keyA] = {reference to valueA}
```

When executing `getObjectByKey`, the cache will behave with the following code:

https://github.com/GalaChain/sdk/blob/main/chaincode/src/types/GalaChainStub.ts#L43-L60

```
  async getCachedState(key: string): Promise<Uint8Array> {
    if (key in this.deletes) {
      return new Uint8Array();
    }

    if (key in this.writes) {
      return this.writes[key];  //  <-- {reference to value identified by string key}
    }

    if (key in this.reads) {
      return this.reads[key];
    }

    const result = await this.stub.getState(key);
    this.reads[key] = result;

    return result;
  }
```

I believe that even a failing transaction can preserve writes to the next Request in the batch,
because `aggregatedCache` and `ctx.stub.reads|writes|deletes` all point to the same values by reference.

When we execute `ctx.stub.setWrites` above, we're still preserving all the same references:

```
  setWrites(writes: Record<string, Uint8Array>): void {
    this.writes = { ...writes };
  }
```

This creates a copy of the top level wrapper, but all nested properties of `...writes` are passed in as references to objects that exist.
Basically, it's not a deep copy, it's a shallow copy.

What we need to be using here, both when we create `aggregatedCache` and when we update it with `ctx.stub.setWrites`,
is something that can create a deep copy. The `class-transformer` method `instanceToInstance`
might be perfect, because it handles deep copies and we're already using these libraries:

https://github.com/typestack/class-transformer?tab=readme-ov-file#instanceToInstance